+++ CHANGELOG.md
@@ -19,6 +19,11 @@ Change categories:
 
 ### Added
 
+- rde: Add NegotiateMediumParameters command
+  - Add `encode_pldm_rde_negotiate_medium_parameters_req()`
+  - Add `decode_pldm_rde_negotiate_medium_parameters_req()`
+  - Add `encode_pldm_rde_negotiate_medium_parameters_resp()`
+  - Add `decode_pldm_rde_negotiate_medium_parameters_resp()`
 - rde: Add NegotiateRedfishParameters command
   - Add `encode_pldm_rde_negotiate_redfish_parameters_req()`
   - Add `decode_pldm_rde_negotiate_redfish_parameters_req()`